requirements.txt
supabase
anthropic
requests
python-dotenv
streamlit
pandas
lob-python
sendgrid
twilio
psycopg2-binary
